requirements.txt
numpy==1.16.1
pandas==0.24.2
pickleshare==0.7.5
Pillow==8.4.0
scipy==1.2.1
scikit-learn==1.0.1
scikit-video==1.1.11
torch==1.10.0
torchaudio==0.10.0
torchvision==0.11.1
